Seperate all secondary classes in Maze.cs to different class files.
